+++ cishouseholds/pipeline/input_file_stages.py
@@ -2,17 +2,6 @@ from cishouseholds.pipeline.mapping import column_name_maps
 from cishouseholds.pipeline.mapping import survey_response_cast_to_double
 from cishouseholds.pipeline.mapping import survey_response_cisd_cast_to_double
 from cishouseholds.pipeline.pipeline_stages import generate_input_processing_function
-from cishouseholds.pipeline.pre_union_transformations import assign_has_been_columns
-from cishouseholds.pipeline.pre_union_transformations import clean_survey_responses_version_1
-from cishouseholds.pipeline.pre_union_transformations import clean_survey_responses_version_2
-from cishouseholds.pipeline.pre_union_transformations import derive_additional_v1_2_columns
-from cishouseholds.pipeline.pre_union_transformations import digital_responses_preprocessing
-from cishouseholds.pipeline.pre_union_transformations import transform_participant_extract_digital
-from cishouseholds.pipeline.pre_union_transformations import transform_survey_responses_version_0_delta
-from cishouseholds.pipeline.pre_union_transformations import transform_survey_responses_version_1_delta
-from cishouseholds.pipeline.pre_union_transformations import transform_survey_responses_version_2_delta
-from cishouseholds.pipeline.pre_union_transformations import transform_survey_responses_version_digital_delta
-from cishouseholds.pipeline.pre_union_transformations import translate_welsh_survey_responses_version_digital
 from cishouseholds.pipeline.timestamp_map import blood_datetime_map
 from cishouseholds.pipeline.timestamp_map import cis_digital_datetime_map
 from cishouseholds.pipeline.timestamp_map import historical_blood_datetime_map
@@ -21,7 +10,27 @@ from cishouseholds.pipeline.timestamp_map import survey_responses_v0_datetime_ma
 from cishouseholds.pipeline.timestamp_map import survey_responses_v1_datetime_map
 from cishouseholds.pipeline.timestamp_map import survey_responses_v2_datetime_map
 from cishouseholds.pipeline.validation_schema import validation_schemas
-
+from cishouseholds.pipeline.version_specific_processing.digital_transformations import digital_responses_preprocessing
+from cishouseholds.pipeline.version_specific_processing.digital_transformations import (
+    transform_survey_responses_version_digital_delta,
+)
+from cishouseholds.pipeline.version_specific_processing.mult_version import assign_has_been_columns
+from cishouseholds.pipeline.version_specific_processing.mult_version import derive_additional_v1_2_columns
+from cishouseholds.pipeline.version_specific_processing.participant_extract import transform_participant_extract_digital
+from cishouseholds.pipeline.version_specific_processing.participant_extract import (
+    translate_welsh_survey_responses_version_digital,
+)
+from cishouseholds.pipeline.version_specific_processing.v0_transformations import (
+    transform_survey_responses_version_0_delta,
+)
+from cishouseholds.pipeline.version_specific_processing.v1_transformations import clean_survey_responses_version_1
+from cishouseholds.pipeline.version_specific_processing.v1_transformations import (
+    transform_survey_responses_version_1_delta,
+)
+from cishouseholds.pipeline.version_specific_processing.v2_transformations import clean_survey_responses_version_2
+from cishouseholds.pipeline.version_specific_processing.v2_transformations import (
+    transform_survey_responses_version_2_delta,
+)
 
 participant_extract_digital_parameters = {
     "stage_name": "participant_extract_digital_ETL",
