+++ cishouseholds/pipeline/visit_transformations.py
@@ -1,7 +1,9 @@
 import pyspark.sql.functions as F
 from pyspark.sql import DataFrame
 
+from cishouseholds.derive import assign_column_given_proportion
 from cishouseholds.derive import assign_date_difference
+from cishouseholds.derive import assign_ever_had_long_term_health_condition_or_disabled
 from cishouseholds.derive import assign_first_visit
 from cishouseholds.derive import assign_last_visit
 from cishouseholds.derive import assign_named_buckets
@@ -10,6 +12,13 @@ from cishouseholds.edit import replace_sample_barcode
 
 
 def visit_transformations(df: DataFrame):
+    """"""
+    df = visit_derivations(df).custom_checkpoint()
+    df = create_ever_variable_columns(df).custom_checkpoint()
+    return df
+
+
+def visit_derivations(df: DataFrame):
 
     df = assign_visit_order(
         df=df,
@@ -55,4 +64,39 @@ def visit_transformations(df: DataFrame):
         column_name_to_assign="visit_number",
         map={**{0: 1, 14: 2, 21: 3, 28: 4}, **{i * 28: (i + 3) for i in range(2, 200)}},
     )
-    pass
+    return df
+
+
+def create_ever_variable_columns(df: DataFrame) -> DataFrame:
+    """"""
+    df = assign_column_given_proportion(
+        df=df,
+        column_name_to_assign="ever_work_person_facing_or_social_care",
+        groupby_column="participant_id",
+        reference_columns=["work_social_care"],
+        count_if=["Yes, care/residential home, resident-facing", "Yes, other social care, resident-facing", "Yes"],
+        true_false_values=["Yes", "No"],
+    )
+    df = assign_column_given_proportion(
+        df=df,
+        column_name_to_assign="ever_care_home_worker",
+        groupby_column="participant_id",
+        reference_columns=["work_social_care", "work_nursing_or_residential_care_home"],
+        count_if=["Yes", "Yes, care/residential home, resident-facing"],
+        true_false_values=["Yes", "No"],
+    )
+    df = assign_column_given_proportion(
+        df=df,
+        column_name_to_assign="ever_had_long_term_health_condition",
+        groupby_column="participant_id",
+        reference_columns=["illness_lasting_over_12_months"],
+        count_if=["Yes"],
+        true_false_values=["Yes", "No"],
+    )
+    df = assign_ever_had_long_term_health_condition_or_disabled(
+        df=df,
+        column_name_to_assign="ever_had_long_term_health_condition_or_disabled",
+        health_conditions_column="illness_lasting_over_12_months",
+        condition_impact_column="illness_reduces_activity_or_ability",
+    )
+    return df
