+++ tests/impute/test_travel_column_imputation.py
@@ -2,7 +2,7 @@ from chispa import assert_df_equality
 
 from cishouseholds.edit import update_to_value_if_any_not_null
 from cishouseholds.impute import fill_forward_from_last_change
-from cishouseholds.pipeline.post_union_transformations import fill_forwards
+from cishouseholds.pipeline.high_level_transformations import fill_forwards
 
 # from cishouseholds.edit import update_travel_column
